+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 - Improve the Langchain::Tool::Database tool
 - Allow explictly setting tool_choice on the Assistant instance
+- Add support for bulk embedding in Ollama
 
 ## [0.15.3] - 2024-08-27
 - Fix OpenAI#embed when text-embedding-ada-002 is used